[Update] label_rotation の型をリファクタリングし、テストカバレッジレポートを更新しました。
- `GsicAddressShape` クラスにおいて、`label_rotation` の型を float から int に変更しました。
- テストカバレッジレポートを更新し、新しいタイムスタンプを追加するとともに、直近のテスト実行状況を反映するようテスト件数を調整しました。
- `compute_visual_offset` が繰り返し呼び出しの際に結果を正しくキャッシュすることを確認するための新しいテストを追加し、パフォーマンスと信頼性を向上させました。

+++ nfj/geospatial.py
@@ -1199,7 +1199,7 @@ class GsicAddressShape(GsShapeFile):
         dxfversion: str = "R2010",
         units: InsertUnits = InsertUnits.Meters,
         label_scale: float = 1.0,
-        label_rotation: float = 0.0,
+        label_rotation: int = 0,
         main_address: bool = True,
         locality: bool = False,
         branch_office: bool = False,
